requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-anthropic>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0
faiss-cpu>=1.8.0
pydantic>=2.0.0
python-dotenv>=1.0.0
rich>=13.0.0
tabulate>=0.9.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
